src/backend/requirements.txt
fastapi
uvicorn
neo4j
chromadb
ollama
pydantic
python-dotenv
requests
pandas
